Abstract out ModelStub from AddCommandTest
